Crash just after app start because of resource directory does not yet exists

+++ C#_Sources/FileManager/StoreSizeForm.cs
@@ -111,7 +111,10 @@ namespace FileManager
 			{
 				formSettignsBody += ((formSettignsBody == "") ? "" : crlf) + newRow;
 			}
-			File.WriteAllText(formSettignsFilePath, formSettignsBody);
+			if (Directory.Exists(getUserSettingsFolderPath()))
+			{
+				File.WriteAllText(formSettignsFilePath, formSettignsBody);
+			}
 		}
 
 		private void restoreSizeAndLocation()
